+++ asyncgit/src/error.rs
@@ -12,6 +12,9 @@ pub enum Error {
     #[error("git: remote url not found")]
     UnknownRemote,
 
+    #[error("git: inconclusive remotes")]
+    NoDefaultRemoteFound,
+
     #[error("git: work dir error")]
     NoWorkDir,