+++ asyncgit/src/sync/mod.rs
@@ -41,7 +41,7 @@ pub use hunks::{reset_hunk, stage_hunk, unstage_hunk};
 pub use ignore::add_to_ignore;
 pub use logwalker::LogWalker;
 pub use remotes::{
-    fetch_origin, get_first_remote, get_remotes, push,
+    fetch_origin, get_default_remote, get_remotes, push,
     ProgressNotification,
 };
 pub use reset::{reset_stage, reset_workdir};